Removed artificial rule for emph/strong markers.
Previously there was a rule that nothing in a string of more than 3
`*` or `_` characters could close or start emphasis. This was artifical
and led to strange asymmetries, e.g. you could have
`*a *b**` emph within emph but not `**a **b****` strong within strong.

The new parsing strategy makes it easy to remove this limitation.

Spec, js, and c implementations have been updated.  Spec might need
some further grooming.