Changed rule for `_` emphasis and strong emphasis.
To prevent intra-word emphasis, we used to check to see if
the delimiter was followed/preceded by an ASCII alphanumeric.

We now do something more elegant:  whereas an opening `*` must
be left-flanking, an opening `_` must be left-flanking *and not
right-flanking*.  And so on for the other cases.

All the original tests passed except some tests with Russian
text with internal `_`, which formerly created emphasis but no
longer do with the new rule.  These tests have been adjusted.
A few new test cases have been added to illustrate the rule.

The C and JS implementations have both been updated.