Rewrote dingus without ACE editor. Closes #11.
* The ACE editor is problematic from an accesibility point of
  view.  See #11.
* The dingus with ACE produces overlapping content in small windows.

What we lose: ability to have soft-wrapped text.  (It's too difficult
to calculate the "real" line number from the scroll position of the
text area if lines can be soft-wrapped.  ACE had a function for
this.)  Perhaps this is for the best, since with soft-wrapped text
the difference between a soft-wrapped and hard-wrapped line is
invisible, and it can matter for rendering.