Fixed minor regex bug with raw HTML blocks.
Closes #98.

This would affect things like:

    <a>[SPACE][SPACE]
    x

which, with the change, gets parsed as a raw HTML
block followed by a paragraph 'x', instead of a single
paragraph with inline HTML, a line brek, and 'x'.

The new behavior conforms to the spec.
See #7 in 4.6.

Perhaps a spec example should be added for this.
Or at least a regression test.