Simplify and optimize brackets processing (links/images)
The key insight for the simplification is that we don't need to keep the
brackets in the same stack as the other delimiters. All we need to know
is which delimiter came before the bracket, so that we can process the
emphasis in the image/link text.

This removes the need to loop to find the last opening bracket and also
removes the need to skip over brackets in processEmphasis.

In addition to that, we can skip the slice and reference lookup for
collapsed/shortcut reference links if we know that it contains a
bracket.

Together, these changes make the "nested brackets 10000 deep"
pathological case go from 400 ms to 20 ms.