Revert "Use esm for bin/commonmark."
This reverts commit 9bcf45b1e384eb0983d023298e0e65bffafceb76.

Node.js version 14, which supports ES Modules without any flag or the
`esm` package, is currently in long-term support. But a great many
folks still run older version of Node.js that either don't support ES
Modules at all or hide that support behind a feature flag.

We could instruct Node.js to load the bin script as an ES Module via
a flag in shebang, but that flag won't be available on older versions
of Node.js. Meanwhile, this bin script doesn't stand to gain anything
from moving to ES Modules. When something runs the bin script, it's
going to be Node.js, not a browser. So we might as well use `require`
and stop worrying about ES Modules.